Revert "adapted to moved AbstractWorkflowConfigurationBuilder"
This reverts commit 8cf7ec1104d480d3a8d0490948829d938eba0b92.
