+++ bundles/de.uka.ipd.sdq.simucom.rerunsimulation/src/de/uka/ipd/sdq/simucom/rerunsimulation/runconfig/RerunSimulationWorkflowLauncher.java
@@ -13,7 +13,7 @@ import de.uka.ipd.sdq.codegen.simucontroller.runconfig.SimuComWorkflowConfigurat
 import de.uka.ipd.sdq.codegen.simucontroller.runconfig.SimuComWorkflowLauncher;
 import de.uka.ipd.sdq.simucom.rerunsimulation.jobs.RerunSimulationJob;
 import de.uka.ipd.sdq.workflow.jobs.IJob;
-import de.uka.ipd.sdq.workflow.launchconfig.core.AbstractWorkflowConfigurationBuilder;
+import de.uka.ipd.sdq.workflow.launchconfig.AbstractWorkflowConfigurationBuilder;
 
 /**
  * This class creates the RerunSimulationJob which is needed for rerunning the simulation without
